requirements.txt
PySide6>=6.7
pyaudio>=0.2.14
numpy>=2.0
torch
openai-whisper
python-docx
requests
pytest
